Adjusting the email validation settings
Until now, the deliverability of an email has also been checked to see
whether the Top Level Domain (TLD) of the address exists. This check fails
if the server is temporarily unavailable, if you are behind a proxy or if
there is no network connection. For this reason, this check is removed.